+++ benches/dowser.rs
@@ -7,10 +7,17 @@ use brunch::{
 	benches,
 };
 use dowser::Dowser;
-use std::time::Duration;
+use std::{
+	path::PathBuf,
+	time::Duration,
+};
 
 benches!(
 	Bench::new("dowser::Dowser", "with_path(/usr/share)")
 		.timed(Duration::from_secs(6))
-		.with(|| Dowser::default().with_path("/usr/share").into_vec()),
+		.with(|| Dowser::default().with_path("/usr/share").collect::<Vec<_>>()),
+
+	Bench::new("dowser::Dowser", "from(/usr/share)")
+		.timed(Duration::from_secs(6))
+		.with(|| Dowser::from(PathBuf::from("/usr/share")).collect::<Vec<_>>()),
 );